refactor : move Member-related files
