+++ server/src/main/java/tig/server/jwt/handler/JwtExceptionHandlerFilter.java
@@ -16,7 +16,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.util.StringUtils;
 import org.springframework.web.filter.OncePerRequestFilter;
 import tig.server.jwt.TokenProvider;
-import tig.server.member.MemberDetailsServiceImpl;
+import tig.server.member.service.MemberDetailsServiceImpl;
 
 import java.io.IOException;
